requirements.txt
catboost==1.0.3
matplotlib==3.4.3
numpy==1.21.4
pandas==1.3.4
scikit-learn==1.0.1
seaborn==0.11.2
ipykernel==6.5.0
sklearn-pandas==2.2.0
scikit-optimize==0.9.0
lightgbm==3.3.1

# Due to an Error occuring:
ipywidgets==7.6.5